+++ changelog.md
@@ -6,6 +6,7 @@
 * [fixed] List of not confirmed now doesn't skip doubled score
 * [fixed] Show remaining identities if one of the suggestion is None
 * [changed] Select representative media files in identification according to the orientation
+* [added] Automatic postponed run of init identification on changed list of representative media files
 
 ## 06-2025